[~] move ApiKeyVerifier, ITokenVerifier and ITokenVerifier into service package
