+++ effectiveOfficeBackend/src/main/kotlin/office/effective/features/simpleAuth/service/RequestVerifier.kt
@@ -5,7 +5,6 @@ import io.ktor.client.engine.cio.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
-import office.effective.features.simpleAuth.ITokenVerifier
 import org.slf4j.LoggerFactory
 
 /**
