+++ effectiveOfficeBackend/src/main/kotlin/office/effective/features/user/facade/UserFacade.kt
@@ -2,7 +2,7 @@ package office.effective.features.user.facade
 
 import office.effective.common.exception.InstanceNotFoundException
 import office.effective.common.utils.DatabaseTransactionManager
-import office.effective.features.simpleAuth.ITokenVerifier
+import office.effective.features.simpleAuth.service.ITokenVerifier
 import office.effective.features.user.converters.UserDTOModelConverter
 import office.effective.dto.UserDTO
 import office.effective.serviceapi.IUserService
